% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{count_SNVmutations}
\alias{count_SNVmutations}
\title{This function returns count table of the SNV mutation types and their frequency, also shows it in the graph}
\usage{
count_SNVmutations(vcf, refgen)
}
\arguments{
\item{vcf}{The VCF file}

\item{refgen}{The reference genome (e.g. BSgenome.Hsapiens.UCSC.hg19)}
}
\value{
A count table with mutation types for each SNV and their frequancy
}
\description{
This function returns count table of the SNV mutation types and their frequency, also shows it in the graph
}
\examples{
library("VariantAnnotation")
library("BSgenome")
library("BSgenome.Hsapiens.UCSC.hg19")
vcffile <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
vcf <- readVcf(vcffile, "hg19")
count_SNVmutations(vcf, Hsapiens)
}
\author{
Aigerim Nugmanova \cr Politecnico di Milano \cr E-Mail: \href{mailto:aigerim.nugmanova@mail.polimi.it}{aigerim.nugmanova@mail.polimi.it}
}
