% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{determine_mutation_types}
\alias{determine_mutation_types}
\title{This function determines the mutation types by returning and constructing a vector of SNV from VCF file, reference genome
and given context length}
\usage{
determine_mutation_types(vcf, refgen, context_len)
}
\arguments{
\item{vcf}{The VCF file}

\item{refgen}{The reference genome (e.g. BSgenome.Hsapiens.UCSC.hg19)}

\item{context_len}{Positive, odd integer specifying the context length}
}
\value{
A vector with mutation types for each SNV in the format UP REF>ALT DOWN
}
\description{
This function determines the mutation types by returning and constructing a vector of SNV from VCF file, reference genome
and given context length
}
\examples{
library("VariantAnnotation")
library("BSgenome")
library("BSgenome.Hsapiens.UCSC.hg19")
vcffile <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
vcf <- readVcf(vcffile, "hg19")
determine_mutation_types(vcf, Hsapiens, 3)
}
\author{
Aigerim Nugmanova \cr Politecnico di Milano \cr E-Mail: \href{mailto:aigerim.nugmanova@mail.polimi.it}{aigerim.nugmanova@mail.polimi.it}
}
