% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekMutAndSeq.R
\name{seekMutAndSeq}
\alias{seekMutAndSeq}
\title{Seek Mutations and Sequences around Single Nucleotide Variants (SNVs)}
\usage{
seekMutAndSeq(df, contextLength = 3, refGen)
}
\arguments{
\item{df}{A data frame containing SNV information, including 'chr', 'pos', 'ref',
'alt', and 'stra' columns.}

\item{contextLength}{The length of the sequence around each SNV. It must be an
odd number greater than or equal to 3. Default is 3.}

\item{refGen}{Reference genome object.}
}
\value{
A data frame with information about mutations and the sequences around
each SNV.
}
\description{
This function takes a dataframe of SNVs and extracts sequences downward and upward
for each SNV based on the provided context length.
It replaces 'A's and 'G's in 'ref'
and 'alt' with 'T' and 'C', retrieves genomic ranges, and fetches sequences
from the reference genome. The resulting data frame includes mutation type
information and the extracted sequences.
}
\details{
The function checks the validity of the input parameters and processes the
data frame to extract sequences around each SNV. It ensures that the sequences
do not go beyond the chromosome range. The resulting data frame includes
mutation type information, where 'A's and 'G's in 'ref' and 'alt' are replaced
with 'T' and 'C', respectively.
}
\examples{
df <- data.frame(
id = c("rs7410291", "rs1234567"),
chr = c("chr22", "chrX"),
pos = c("50300078", "100200300"),
start = c("0", "0"),
end = c("0", "0"),
ref = c("A", "C"),
alt = c("G", "T"),
stra = c("*", "*")
)
refgenome <- BSgenome.Hsapiens.UCSC.hg38
seekMutAndSeq(df, contextLength=3, refgenome)

}
