% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekSNVs.R
\name{seekSNVs}
\alias{seekSNVs}
\title{Seek Single Nucleotide Variants (SNVs) from a VCF file}
\usage{
seekSNVs(vcfFile, referenceGenome)
}
\arguments{
\item{vcfFile}{Path to the VCF file or an object of class "CollapsedVCF".}

\item{referenceGenome}{Character string specifying the reference genome.}
}
\value{
A dataframe with SNV information, including ID, chromosome, position,
reference base, alternative base, and strand information.
}
\description{
This function extracts SNVs from a given VCF file and returns a dataframe
with necessary information for further analysis.
}
\details{
The function first checks the input parameters and then processes the VCF file
to extract SNVs. SNVs are defined as variants where both the reference and
alternative alleles have a length of 1 according to the Variant Classification
definition (https://genome.sph.umich.edu/wiki/Variant_classification).
}
\examples{
vcfdata <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
results <- seekSNVs(vcfdata, referenceGenome="hg19")

}
